pyproject.toml
[tool.poetry]
name = "ska-mid-cbf-common-test-infrastructure"
version = "0.0.0"
description = ""
authors = [
    "Rob Huxtable <robert.huxtable@mda.space>",
    "Anton Deriabine <anton.deriabine@mda.space>",
    "Elizabeth Pieters <elizabeth.pieters@mda.space>",
    "Parmis Mohaghegh <parmis.mohaghegh@mda.space>",
    "Dave Kaufman <dave.kaufman@mda.space>",
    "Brandon Dos Remedios <brandon.dosremedios@mda.space>",
]
license = "BSD License"
readme = "README.md"

packages = [
    { include = "ska_mid_cbf_common_test_infrastructure", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
isort = "^5.6.4"
flake8 = "^3.9.2"
black = "^22.3.0"
jsonschema = "^4.18.4"
pytest-json-report = "^1.5.0"
pytest-cov = "^4.1.0"
pylint = "^2.17.4"
pylint-junit = "^0.3.2"
pytest-json = "^0.4.0"
pylance = "^0.5.9"
pytango = "9.5.0"
ska-tango-base = "1.0.0"
ska-tango-testing = "0.7.2"

[[tool.poetry.source]]
name = "nexus-internal"
url = "https://artefact.skao.int/repository/pypi-internal/simple"

[[tool.poetry.source]]
name = "PyPI-public"
url = "https://pypi.org/simple"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.4.7"
sphinx-rtd-theme = "^3.0.0"
myst-parser = "^4.0.1"
pytango = "9.5.0"
ska-tango-base = "1.0.0"
ska-tango-testing = "0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
